WIP: run a test for each of the ringsizes 16-512
This means we can explore larger clusters. The original test is capped
at 15 nodes since it could generate a ring of 16, or 512, or any
inbetween. With the state space so small why not generate each of the
rings from 16-512 and for each ring size allow up to RingSize nodes to
be added. Looking at test runs it seems we never get more than ~50 nodes
even for 512, and there is some tweaking to be done, but this is still
better coverage than the original tests.